Add resampling and interpolation functions for log data
- Implement linear interpolation, circular interpolation (degrees and radians).
- Create resampleData function to resample log records at specified frequencies.
- Add tests for interpolation functions and resampling logic.
- Define LogRecord and related types in a new types file.
- Update filters and log store to use new types.
- Remove unused GPS parsing utility and related tests.
- Clean up unused utility functions and tests related to path color.
