+++ src/store/filters.ts
@@ -1,6 +1,6 @@
 import { create } from 'zustand'
 import { FiltersType, type FilterLogItem } from '@/types/filters.ts'
-import type { RadiomasterLogRecord } from '@/types/data.ts'
+import type { RadiomasterLogRecord } from '@/parse/types'
 
 export type FiltersData = Record<FiltersType, FilterLogItem<any>>
