--- src/utils/gps/index.ts
@@ -1,10 +1,0 @@
-export function parseGpsLocation(
-  gps: string,
-): { lat: number; lng: number } | null {
-  if (!gps) return null
-
-  const [lat, lng] = gps.split(' ').map(parseFloat)
-  if (isNaN(lat) || isNaN(lng)) return null
-
-  return { lat, lng }
-}
