--- src/utils/parse/parse.test.ts
@@ -1,28 +1,0 @@
-import { describe, it, expect } from 'vitest'
-import { parseCsv } from './'
-
-describe('Parse CSV', () => {
-  it('should parse CSV text into records', async () => {
-    const csvText = `name,age` + `\nAlice,30` + `\nBob,25`
-
-    const expectedRecords = [
-      { name: 'Alice', age: '30' },
-      { name: 'Bob', age: '25' },
-    ]
-
-    const records = await parseCsv(csvText)
-    expect(records).toEqual(expectedRecords)
-  })
-
-  it('should handle empty lines', async () => {
-    const csvText = `name,age` + `\nAlice,30` + `\n` + `\nBob,25`
-
-    const expectedRecords = [
-      { name: 'Alice', age: '30' },
-      { name: 'Bob', age: '25' },
-    ]
-
-    const records = await parseCsv(csvText)
-    expect(records).toEqual(expectedRecords)
-  })
-})
